DRTII-602 rename package and UTC+2 timezone for Departure time
