--- src/main/scala/uk/gov/homeoffice/Main.scala
@@ -1,13 +1,0 @@
-package uk.gov.homeoffice
-
-import cats.effect.{ExitCode, IO, IOApp}
-import uk.gov.homeoffice.AppEnvironment.config
-
-object Main extends IOApp {
-  def run(args: List[String]) =
-
-    config.load[IO].flatMap { cfg =>
-      IEIDashbordServer.stream[IO](cfg).compile.drain.as(ExitCode.Success)
-    }
-
-}
