+++ src/main/scala/uk/gov/homeoffice/drt/drt/applicative/ArrivalFlights.scala
@@ -1,0 +1,25 @@
+package uk.gov.homeoffice.drt.drt.applicative
+
+import cats.Applicative
+import cats.syntax.all._
+import uk.gov.homeoffice.drt.drt.model.{Arrivals, RequestedFlightDetails}
+import uk.gov.homeoffice.drt.drt.repository.ArrivalTableData
+import uk.gov.homeoffice.drt.drt.service.ArrivalService
+
+
+trait ArrivalFlights[F[_]] {
+  def flights(n: RequestedFlightDetails): F[Arrivals]
+}
+
+object ArrivalFlights {
+  implicit def apply[F[_]](implicit ev: ArrivalFlights[F]): ArrivalFlights[F] = ev
+
+  def impl[F[_] : Applicative](arrivalsService: ArrivalService[F]): ArrivalFlights[F] = new ArrivalFlights[F] {
+    def flights(n: RequestedFlightDetails): F[Arrivals] = {
+      val a: F[List[(ArrivalTableData, Int)]] = arrivalsService.getFlightsDetail(n)
+      arrivalsService.transformArrivals(a).map(Arrivals(_))
+    }
+  }
+}
+
+
