+++ src/main/scala/uk/gov/homeoffice/drt/drt/model/Arrival.scala
@@ -1,0 +1,13 @@
+package uk.gov.homeoffice.drt.drt.model
+
+import java.util.Date
+
+case class Arrivals(data: List[Arrival])
+
+case class Arrival(_id: String,
+                   scheduledArrivalDate: Date,
+                   carrierName: String,
+                   flightNumber: String,
+                   arrivingAirport: String,
+                   origin: String,
+                   scheduledDepartureTime: Option[Date])
