+++ src/main/scala/uk/gov/homeoffice/drt/drt/model/RequestedFlightDetails.scala
@@ -1,0 +1,4 @@
+package uk.gov.homeoffice.drt.drt.model
+
+final case class RequestedFlightDetails(region: String, country: String, date: String)
+
