+++ src/main/scala/uk/gov/homeoffice/drt/drt/utils/FileUtil.scala
@@ -1,0 +1,11 @@
+package uk.gov.homeoffice.drt.drt.utils
+
+import scala.io.Source
+
+object FileUtil {
+
+  def readResourceFile(path: String) = {
+    Source.fromResource(path).mkString
+  }
+
+}
