--- src/main/scala/uk/gov/homeoffice/model/RequestedFlightDetails.scala
@@ -1,4 +1,0 @@
-package uk.gov.homeoffice.model
-
-final case class RequestedFlightDetails(region: String, country: String, date: String)
-
