--- src/main/scala/uk/gov/homeoffice/utils/FileUtil.scala
@@ -1,11 +1,0 @@
-package uk.gov.homeoffice.utils
-
-import scala.io.Source
-
-object FileUtil {
-
-  def readResourceFile(path: String) = {
-    Source.fromResource(path).mkString
-  }
-
-}
